Started adding tests
